// generated/src/app/api-models/cap-cputl-rpt-wrk.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapCputlRptWrk {
  seqCapCputlRptWrk: number;
  seqJobId: number;
  capModelId: string;
  provId: string;
  provLastName: string;
  provFirstName: string;
  provMiddleInitial: string;
  provProviderType: string;
  provPrimarySpecialtyType: string;
  vendId: string;
  vendFullName: string;
  groupId: string;
  groupName: string;
  capPoolId: string;
  medDefCode: string;
  medDefDesc: string;
  capCreditAmt: number;
  capMedCostAmt: number;
  capStoplossAmt: number;
  capOtherAdjAmt: number;
  capWithholdAmt: number;
  capCapFfsAmt: number;
  capUnitCount: number;
  capModelPmpm: number;
  capPcpPmpm: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  adjustmentFlag: string;
}